Move blacklist persistence code to redux enhancer
- as per discussion in PR #93
