+++ src/options/blacklist/index.js
@@ -1,6 +1,6 @@
 import * as actions from './actions'
 import * as constants from './constants'
 export { default as reducer } from './reducers'
-export { default as configurePersistence } from './persistence'
+export { default as enhancer } from './enhancer'
 export { default } from './container'
 export { actions, constants }
