+++ src/options/store.js
@@ -1,15 +1,25 @@
-import { createStore, combineReducers } from 'redux'
+import { createStore, combineReducers, compose } from 'redux'
 
-import { reducer as blacklist } from './blacklist'
+import * as blacklist from './blacklist'
 
 const rootReducer = combineReducers({
-    blacklist,
+    blacklist: blacklist.reducer,
 })
 
-export default function configureStore({ReduxDevTools=undefined}={}) {
+export default function configureStore({ReduxDevTools = undefined} = {}) {
+    const enhancers = [
+        blacklist.enhancer,
+    ]
+
+    if (ReduxDevTools) {
+        enhancers.push(ReduxDevTools.instrument())
+    }
+
+    const enhancer = compose(...enhancers)
+
     return createStore(
         rootReducer,
         undefined, // Initial State
-        ReduxDevTools && ReduxDevTools.instrument()
+        enhancer
     )
 }
